move landing page to the root and update header to link to the landing page and remove old link to lp
