+++ app/page.js
@@ -1,21 +1,28 @@
 "use client";
-
+import { useContext } from "react";
+import Link from "next/link";
 import styles from "./page.module.css";
-import { useContext, useEffect } from "react";
+import { AuthContext } from "@/app/context/authContext";
 
-import { ContentContext } from "./context/contentContext";
-
-import ProtectedRoute from "./routes/ProtectedRoute";
-import FormsSection from "./components/FormsSections/FormsSection";
-
-export default function Home() {
-  const { loading } = useContext(ContentContext);
-
-  const screenToDisplay = loading ? <div>Loading...</div> : <FormsSection />;
-
+export default function LandingPage() {
+  const { isAuthenticated } = useContext(AuthContext);
   return (
-    <main className={styles.main}>
-      <ProtectedRoute>{screenToDisplay}</ProtectedRoute>
-    </main>
+    <div className={styles.container}>
+      <section className={styles.heroMessage}>
+        <h1>OWN YOUR LEARNING JOURNEY</h1>
+        <p className={styles.description}>
+          Learning fast doesn’t mean rushing. <br />
+          Pause. Reflect. Plan. Act.
+          <br />
+          Take control of your progress with intention.
+        </p>
+        {!isAuthenticated ? (
+          <Link href="/routes/auth">
+            <button type="button">Start your Learning Audit now</button>
+          </Link>
+        ) : null}
+      </section>
+      <section aria-label="Screenshots section" className={styles.screenshots}></section>
+    </div>
   );
 }
