--- app/routes/landingPage/landingPage.module.css
@@ -1,27 +1,0 @@
-.container {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  padding: 1rem;
-}
-
-.heroMessage {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-  gap: 2rem;
-}
-
-.heroMessage h1 {
-  font-size: 5rem;
-  letter-spacing: 0.25rem;
-}
-
-.heroMessage p {
-  font-size: 1.25rem;
-}
-
-.heroMessage button {
-  font-size: 1.5rem;
-}
